Remove two-way sync between Rigid_body components and the underlying data store. Now, changes to the component data after initialisation won't flow through.
Introduced a sync_from_store system that writes basic pos and vel data
back to the component, so it can act as a snapshot.

This was done to reduce the complexity of the process. Allowing the
component data to be directly mutable was confusing and error prone.
Kinematic type bodies will need a way to be updated by game code soon
though.